+++ test/configure/updatePackageManifest.test.js
@@ -1,0 +1,75 @@
+//
+// Copyright (c) Microsoft Corporation
+//
+// This source code is licensed under the MIT license found in the
+// LICENSE file in the root directory of this source tree.
+//
+// @ts-check
+
+jest.mock("fs");
+
+describe("updatePackageManifest()", () => {
+  const { mockFiles } = require("./mockFiles");
+  const { updatePackageManifest } = require("../../scripts/configure");
+
+  afterEach(() => mockFiles());
+
+  test("adds `scripts` field if missing", () => {
+    mockFiles(["package.json", { key: "value" }]);
+
+    const config = {
+      scripts: {
+        test: "jest",
+      },
+      dependencies: {},
+      files: {},
+      oldFiles: [],
+    };
+
+    expect(updatePackageManifest("package.json", config)).toEqual({
+      key: "value",
+      scripts: {
+        test: "jest",
+      },
+      dependencies: {},
+      devDependencies: {
+        mkdirp: "^1.0.0",
+        "react-native-test-app": "^0.0.1-dev",
+      },
+    });
+  });
+
+  test("adds to existing `scripts` field", () => {
+    mockFiles([
+      "package.json",
+      {
+        key: "value",
+        scripts: {
+          test: "jest",
+        },
+      },
+    ]);
+
+    const config = {
+      scripts: {
+        run: "run",
+      },
+      dependencies: {},
+      files: {},
+      oldFiles: [],
+    };
+
+    expect(updatePackageManifest("package.json", config)).toEqual({
+      key: "value",
+      scripts: {
+        run: "run",
+        test: "jest",
+      },
+      dependencies: {},
+      devDependencies: {
+        mkdirp: "^1.0.0",
+        "react-native-test-app": "^0.0.1-dev",
+      },
+    });
+  });
+});
